# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "uiscloud_weknora 지식베이스 관리 시스템 API 문서",
        "title": "uiscloud_weknora API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "uiscloud_weknora Github",
            "url": "https://github.com/rockgis/uiscloud_weknora"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/chunks/by-id/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "청크 ID만으로 청크 상세 정보를 조회합니다 (knowledge_id 불필요)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "청크 관리"
                ],
                "summary": "ID로 청크 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "청크 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "청크 상세 정보",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 파라미터",
                        "schema": {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_errors.AppError"
                        }
                    },
                    "404": {
                        "description": "청크가 존재하지 않습니다",
                        "schema": {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_errors.AppError"
                        }
                    }
                }
            }
        },
        "/chunks/by-id/{id}/questions": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "청크에서 생성된 질문 삭제",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "청크 관리"
                ],
                "summary": "생성된 질문 삭제",
                "parameters": [
                    {
                        "type": "string",
                        "description": "청크 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "질문 ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "question_id": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "삭제 성공",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 파라미터",
                        "schema": {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_errors.AppError"
                        }
                    },
                    "404": {
                        "description": "청크가 존재하지 않습니다",
                        "schema": {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_errors.AppError"
                        }
                    }
                }
            }
        },
        "/chunks/{knowledge_id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "지정된 지식의 모든 청크 목록을 조회합니다. 페이지네이션을 지원합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "청크 관리"
                ],
                "summary": "지식 청크 목록 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "지식 ID",
                        "name": "knowledge_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "페이지 번호",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "페이지당 항목 수",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "청크 목록",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 파라미터",
                        "schema": {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_errors.AppError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "지정된 지식의 모든 청크를 삭제합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "청크 관리"
                ],
                "summary": "지식 하의 모든 청크 삭제",
                "parameters": [
                    {
                        "type": "string",
                        "description": "지식 ID",
                        "name": "knowledge_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "삭제 성공",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 파라미터",
                        "schema": {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_errors.AppError"
                        }
                    }
                }
            }
        },
        "/chunks/{knowledge_id}/{id}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "지정된 청크의 내용 및 속성을 업데이트합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "청크 관리"
                ],
                "summary": "청크 업데이트",
                "parameters": [
                    {
                        "type": "string",
                        "description": "지식 ID",
                        "name": "knowledge_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "청크 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "업데이트 요청",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.UpdateChunkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "업데이트된 청크",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 파라미터",
                        "schema": {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_errors.AppError"
                        }
                    },
                    "404": {
                        "description": "청크가 존재하지 않습니다",
                        "schema": {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_errors.AppError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "지정된 청크를 삭제합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "청크 관리"
                ],
                "summary": "청크 삭제",
                "parameters": [
                    {
                        "type": "string",
                        "description": "지식 ID",
                        "name": "knowledge_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "청크 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "삭제 성공",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 파라미터",
                        "schema": {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_errors.AppError"
                        }
                    },
                    "404": {
                        "description": "청크가 존재하지 않습니다",
                        "schema": {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_errors.AppError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_Tencent_uiscloud_weknora_internal_errors.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_errors.ErrorCode"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_Tencent_uiscloud_weknora_internal_errors.ErrorCode": {
            "type": "integer",
            "enum": [
                1000,
                1001,
                1002,
                1003,
                1004,
                1005,
                1006,
                1007,
                1008,
                1009,
                1010,
                2000,
                2001,
                2002,
                2003,
                2004,
                2100,
                2101,
                2102,
                2103
            ],
            "x-enum-varnames": [
                "ErrBadRequest",
                "ErrUnauthorized",
                "ErrForbidden",
                "ErrNotFound",
                "ErrMethodNotAllowed",
                "ErrConflict",
                "ErrTooManyRequests",
                "ErrInternalServer",
                "ErrServiceUnavailable",
                "ErrTimeout",
                "ErrValidation",
                "ErrTenantNotFound",
                "ErrTenantAlreadyExists",
                "ErrTenantInactive",
                "ErrTenantNameRequired",
                "ErrTenantInvalidStatus",
                "ErrAgentMissingThinkingModel",
                "ErrAgentMissingAllowedTools",
                "ErrAgentInvalidMaxIterations",
                "ErrAgentInvalidTemperature"
            ]
        },
...
        "internal_handler.UpdateChunkRequest": {
            "type": "object",
            "properties": {
                "chunk_index": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "embedding": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "end_at": {
                    "type": "integer"
                },
                "image_info": {
                    "type": "string"
                },
                "is_enabled": {
                    "type": "boolean"
                },
                "start_at": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Bearer {token} 형식의 JWT 토큰을 입력하세요",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}